Fix bugs and initiate tests

+++ src/main/java/nature/calculadoraco2/Repositories/AtividadedoUsuarioRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AtividadedoUsuarioRepository extends JpaRepository<AtividadedoUsuario, Integer> {
+public interface AtividadedoUsuarioRepository extends JpaRepository<AtividadedoUsuario, Long> {
+    // Métodos adicionais se necessário
 }
